Add files via upload
An analysis was undertaken employing Tableau to scrutinize trends in Surgical Site Infections (SSI) rates, with a particular emphasis on the impact of varying different medical professions and departments on infection outcomes. Through the visualization of data, patterns were discerned, highlighting disparities in SSI rates attributable to professional involvement. This facilitated the identification of critical areas for intervention, thereby contributing to the optimization of infection control protocols and the enhancement of patient care methodologies.